requirements.txt
Django==1.9.6
django-filter==0.13.0
django-rest-framework-proxy==1.6.0
djangorestframework==3.3.3
Pillow==3.2.0
python-dateutil==2.5.3
requests==2.10.0
six==1.10.0
